[REFACTOR] Move InvestigatorAttributes InvestigatorCharacteristics to helpers
